requirements.txt
torch~=1.8.1
PyYAML==5.4.1
tqdm
scikit-learn